Sync with proposed Calico/Liberty charm-helpers code
This is the Calico/Liberty support that we have a current merge
proposal for at
https://code.launchpad.net/~project-calico/charm-helpers/liberty/+merge/287927:

"Add support for Calico networking with OpenStack Liberty release.

"Specifically, with Liberty we use a Calico-specific DHCP agent
instead of the Neutron reference DHCP agent. Therefore we install an
additional package, calico-dhcp-agent, and manage the
'calico-dhcp-agent' service instead of 'neutron-dhcp-agent'."